+++ CHANGES.md
@@ -1,5 +1,8 @@
 Poodinis Changelog
 ==================
+**Version NEXT**
+* ADD setting persistent registration options
+
 **Version 6.0.0**
 * CHANGE registration scopes are replaced by a single factory implementation. If you were not doing anything with the internal scope mechanism, you 
 should not be affected by this change.